+++ Constants.py
@@ -1,7 +1,7 @@
 # Created by moeheart at 10/11/2020
 # 全局用到的部分常量。
 
-EDITION = "7.9.1"
+EDITION = "7.10.0beta"
 ANNOUNCEMENT = "五奶复盘已经全部写好了！接下来就是建设生态环境的时候了~"
 
 LVLTABLE = [0, 30, 100, 200, 300, 500, 700, 900, 1200, 1500, 1800, 2400, 3000, 999999]
