+++ update.md
@@ -6,29 +6,33 @@
 
 # (TODO)7.10.0
 
-(TODO)重构评分系统，现在会用固定化的流程去维护评分结果。
+- (TODO)重构评分系统为"天网系统"，现在会用固定化的流程去维护评分结果。
+  - (TODO)在本地获取数据并预览
+  - (TODO)自动化获取所有数据
+  - (TODO)服务器端获取并存到数据库
+  - (TODO)本地获取并独立设置文件
+  - (TODO)本地展示
+  - (TODO)服务器展示
 
-- (TODO)自动化批量统计数值并保存
+- (TODO)所有复盘支持网页版显示。
+  - (TODO)展示UI模块化
+  - (TODO)联动天网系统
+  - (TODO)本地实现
+  - (TODO)服务器实现
 
-- (TODO)增加读取与发布数值的接口
+- (TODO)网页版时间轴优化
+   - (TODO)增加悬浮窗详细信息
+   - (TODO)时间轴增加比例缩放
+   - (TODO)为时间轴增加一个说明
 
-- (TODO)与网页版的联动
-
-(TODO)所有复盘支持网页版显示。
-
-- (TODO)技能复盘结果格式化生成
-
-- (TODO)技能复盘预留详情接口
-
-- (TODO)时间轴增加悬浮窗详细信息
-
-- (TODO)时间轴增加比例缩放
-
-- (TODO)时间轴增加一个说明
-
-(TODO)为网页版复盘提供一个大厅，可以看到上传的所有公开复盘。
-
-(TODO)数据已被他人上传时提供一个链接。
+- (TODO)完善天网系统
+  - (TODO)为网页版复盘提供一个大厅，可以看到上传的所有公开复盘。
+  - (TODO)数据已被他人上传时提供结果的展示链接。
+    
+- (TODO)新增"专案组"模块，对标"手法警察"，对数据进行分析并且指出其中的优化空间。
+  - (TODO)主逻辑
+  - (TODO)本地展示
+  - (TODO)网页展示
 
 ## 修复
